feat: Improve CLI usage
This feature is a refactor.
The previous logic module is split
into multiple new modules.

Duties are not loaded in a global variable
anymore, but rather in a collection.

The underlying dependency, failprint,
was also updated to provide a function
that allows to reuse its CLI options.

Each duty on the CLI can now accept
flags to alter its context options
(quiet, silent, capture, etc.).